Integrate spdlog-based SpdLogger
Added SpdLogger implementation using spdlog, supporting unified logging, ImGui panel integration. Updated PlayerScript to use new SPD_DEBUG macros. Refactored SpdLogger header to hide spdlog dependencies and moved internal sink implementation to a private header. Updated project files to include new sources and headers.
