+++ GameCode/PlayerScript.hpp
@@ -10,7 +10,7 @@
 #include <sstream>
 #include <vector>
 #include <Math/Vec3.hpp>
-
+#include <Core/SpdLogger.hpp>
 
 /**
  * Example player script demonstrating how to implement IScript.
@@ -26,16 +26,20 @@ public:
         REGISTER_FIELD(godMode);
         REGISTER_FIELD(label);
 
-        LogMessage("PlayerScript created");
+        //LogMessage("PlayerScript created");
+		SPD_DEBUG("PlayerScript created");
+
     }
     
     ~PlayerScript() override {
-        LogMessage("PlayerScript destroyed");
+        //LogMessage("PlayerScript destroyed");
+		SPD_DEBUG("PlayerScript destroyed");
     }
 
     // === IScript Interface ===
     void Initialize(NE::ECS::Entity entity) override {
-        LogMessage("PlayerScript initialized for entity " + std::to_string(entity));
+        //LogMessage("PlayerScript initialized for entity " + std::to_string(entity));
+		SPD_DEBUG("PlayerScript initialized for entity {}", entity);
 
         // In a real implementation, you might:
         // - Get references to other components (Transform, Renderer, etc.)
@@ -48,8 +52,9 @@ public:
         m_timeSinceLastLog += deltaTime;
         
         if (m_timeSinceLastLog >= LOG_INTERVAL) {
-            LogMessage("PlayerScript updating - Entity: " + std::to_string(GetEntity()) + 
-                      ", DeltaTime: " + std::to_string(deltaTime));
+            //LogMessage("PlayerScript updating - Entity: " + std::to_string(GetEntity()) + 
+            //          ", DeltaTime: " + std::to_string(deltaTime));
+			SPD_DEBUG("PlayerScript updating - Entity: {}, DeltaTime: {}", GetEntity(), deltaTime);
             m_timeSinceLastLog = 0.0;
         }
 
@@ -70,7 +75,8 @@ public:
     }
 
     void OnDestroy() override {
-        LogMessage("PlayerScript cleanup for entity " + std::to_string(GetEntity()));
+        //LogMessage("PlayerScript cleanup for entity " + std::to_string(GetEntity()));
+		SPD_DEBUG("PlayerScript cleanup for entity {}", GetEntity());
 
         // In a real implementation, you might:
         // - Unsubscribe from events
@@ -79,12 +85,14 @@ public:
     }
 
     void OnEnable() override {
-        LogMessage("PlayerScript enabled for entity " + std::to_string(GetEntity()));
+        //LogMessage("PlayerScript enabled for entity " + std::to_string(GetEntity()));
+		SPD_DEBUG("PlayerScript enabled for entity {}", GetEntity());
         // Resume player functionality
     }
 
     void OnDisable() override {
-        LogMessage("PlayerScript disabled for entity " + std::to_string(GetEntity()));
+        //LogMessage("PlayerScript disabled for entity " + std::to_string(GetEntity()));
+		SPD_DEBUG("PlayerScript disabled for entity {}", GetEntity());
         // Pause player functionality, reset input
     }
 
@@ -94,20 +102,24 @@ public:
 
     // === Event Handlers ===
     void OnCollisionEnter(NE::ECS::Entity other) override {
-        LogMessage("PlayerScript collision enter with entity " + std::to_string(other));
+        //LogMessage("PlayerScript collision enter with entity " + std::to_string(other));
+		SPD_DEBUG("PlayerScript collision enter with entity {}", other);
         // Your collision logic goes here
     }
 
     void OnCollisionExit(NE::ECS::Entity other) override {
-        LogMessage("PlayerScript collision exit with entity " + std::to_string(other));
+        //LogMessage("PlayerScript collision exit with entity " + std::to_string(other));
+		SPD_DEBUG("PlayerScript collision exit with entity {}", other);
     }
 
     void OnTriggerEnter(NE::ECS::Entity other) override {
         LogMessage("PlayerScript trigger enter with entity " + std::to_string(other));
+		SPD_DEBUG("PlayerScript trigger enter with entity {}", other);
     }
 
     void OnTriggerExit(NE::ECS::Entity other) override {
-        LogMessage("PlayerScript trigger exit with entity " + std::to_string(other));
+        //LogMessage("PlayerScript trigger exit with entity " + std::to_string(other));
+		SPD_DEBUG("PlayerScript trigger exit with entity {}", other);
     }
 
     // === Exposed editable fields via registry === NEED TO CHANGE
